Move config context helpers into their own module
