+++ packages/client-library-otel/src/config/context/index.ts
@@ -1,0 +1,5 @@
+/** Module for managing the resolved config in the OpenTelemetry context */
+
+export { getFpResolvedConfig, setFpResolvedConfig } from "./context";
+export { getRedactedHeaders } from "./redacted-headers";
+export { getRedactedQueryParams } from "./redacted-query-params";
