+++ packages/client-library-otel/src/config/context/redacted-query-params.ts
@@ -1,15 +1,14 @@
 import {
-  DEFAULT_REDACTED_HEADERS,
+  DEFAULT_REDACTED_QUERY_PARAMS,
   type FpResolvedConfig,
-  getFpResolvedConfig,
 } from "../config";
-
+import { getFpResolvedConfig } from "./context";
 /**
  * Helper function to determine which headers to redact
  */
-export function getRedactedHeaders(
+export function getRedactedQueryParams(
   resolvedConfig?: FpResolvedConfig,
 ): Set<string> {
   const config = resolvedConfig ?? getFpResolvedConfig();
-  return config?.redactedHeaders ?? new Set(DEFAULT_REDACTED_HEADERS);
+  return config?.redactedQueryParams ?? new Set(DEFAULT_REDACTED_QUERY_PARAMS);
 }
